Move code from sbt 1 folders to main folder
