+++ dev-mode/sbt-plugin/src/main/scala/play/sbt/PlayImport.scala
@@ -11,7 +11,7 @@ import play.dev.filewatch.FileWatchService
 /**
  * Declares the default imports for Play plugins.
  */
-object PlayImport extends PlayImportCompat {
+object PlayImport {
   val Production = config("production")
 
   def component(id: String) = "com.typesafe.play" %% id % play.core.PlayVersion.current
