+++ dev-mode/sbt-plugin/src/main/scala/play/sbt/routes/RoutesCompiler.scala
@@ -4,6 +4,7 @@
 
 package play.sbt.routes
 
+import play.api.PlayException
 import play.core.PlayVersion
 import play.routes.compiler.RoutesGenerator
 import play.routes.compiler.RoutesCompilationError
@@ -13,9 +14,13 @@ import Compiler.GeneratedSource
 
 import sbt._
 import sbt.Keys._
-import com.typesafe.sbt.web.incremental._
-import play.api.PlayException
 import sbt.plugins.JvmPlugin
+
+import xsbti.Position
+
+import java.util.Optional
+import scala.collection.mutable
+import com.typesafe.sbt.web.incremental._
 
 import scala.language.implicitConversions
 
@@ -54,7 +59,7 @@ object RoutesKeys {
   val InjectedRoutesGenerator = play.routes.compiler.InjectedRoutesGenerator
 }
 
-object RoutesCompiler extends AutoPlugin with RoutesCompilerCompat {
+object RoutesCompiler extends AutoPlugin {
   import RoutesKeys._
 
   override def trigger = noTrigger
@@ -137,6 +142,44 @@ object RoutesCompiler extends AutoPlugin with RoutesCompilerCompat {
     routesGenerator := InjectedRoutesGenerator,
     sourcePositionMappers += routesPositionMapper
   )
+
+  private val routesPositionMapper: Position => Option[Position] = position => {
+    position.sourceFile.asScala.collect {
+      case GeneratedSource(generatedSource) => {
+        new xsbti.Position {
+          override lazy val line: Optional[Integer] = {
+            position.line.asScala
+              .flatMap(l => generatedSource.mapLine(l.asInstanceOf[Int]))
+              .map(l => l.asInstanceOf[java.lang.Integer])
+              .asJava
+          }
+          override lazy val lineContent: String = {
+            line.asScala
+              .flatMap { lineNumber =>
+                sourceFile.asScala.flatMap { file =>
+                  IO.read(file).split('\n').lift(lineNumber - 1)
+                }
+              }
+              .getOrElse("")
+          }
+          override val offset: Optional[Integer]      = Optional.empty[java.lang.Integer]
+          override val pointer: Optional[Integer]     = Optional.empty[java.lang.Integer]
+          override val pointerSpace: Optional[String] = Optional.empty[String]
+          override val sourceFile: Optional[File]     = Optional.ofNullable(generatedSource.source.get)
+          override val sourcePath: Optional[String]   = Optional.ofNullable(sourceFile.get.getCanonicalPath)
+          override lazy val toString: String = {
+            val sb = new mutable.StringBuilder()
+
+            if (sourcePath.isPresent) sb.append(sourcePath.get)
+            if (line.isPresent) sb.append(":").append(line.get)
+            if (lineContent.nonEmpty) sb.append("\n").append(lineContent)
+
+            sb.toString()
+          }
+        }
+      }
+    }
+  }
 
   private val compileRoutesFiles = Def.task[Seq[File]] {
     val log = state.value.log
