+++ dev-mode/sbt-plugin/src/main/scala/play/sbt/run/PlayRun.scala
@@ -28,10 +28,13 @@ import com.typesafe.sbt.packager.universal.UniversalPlugin.autoImport._
 import com.typesafe.sbt.packager.Keys.executableScriptName
 import com.typesafe.sbt.web.SbtWeb.autoImport._
 
+import sbt.internal.io.PlaySource
+import scala.sys.process._
+
 /**
  * Provides mechanisms for running a Play application in sbt
  */
-object PlayRun extends PlayRunCompat {
+object PlayRun {
   class TwirlSourceMapping extends GeneratedSourceMapping {
     def getOriginalLine(generatedSource: File, line: Integer): Integer = {
       MaybeGeneratedSource.unapply(generatedSource).map(_.mapLine(line): java.lang.Integer).orNull
@@ -173,6 +176,59 @@ object PlayRun extends PlayRunCompat {
       twiddleRunMonitor(watched, newState, reloader, Some(newWatchState))
     } else {
       ()
+    }
+  }
+
+  private def sleepForPoolDelay = Thread.sleep(Watched.PollDelay.toMillis)
+
+  private def getPollInterval(watched: Watched): Int = watched.pollInterval.toMillis.toInt
+
+  private def getSourcesFinder(watched: Watched, state: State): PlaySourceModificationWatch.PathFinder = () => {
+    watched
+      .watchSources(state)
+      .map(source => new PlaySource(source))
+      .flatMap(_.getFiles)
+      .collect {
+        case f if f.exists() => better.files.File(f.toPath)
+      }(scala.collection.breakOut)
+  }
+
+  private def kill(pid: String) = s"kill -15 $pid".!
+
+  private def createAndRunProcess(args: Seq[String]) = args.!
+
+  private def watchContinuously(state: State, sbtVersion: String): Option[Watched] = {
+    // sbt 1.1.5+ uses Watched.ContinuousEventMonitor while watching the file system.
+    def watchUsingEvenMonitor = {
+      // If we have Watched.ContinuousEventMonitor attribute and its state.count
+      // is > 0 then we assume we're in ~ run mode
+      state
+        .get(Watched.ContinuousEventMonitor)
+        .map(_.state())
+        .filter(_.count > 0)
+        .flatMap(_ => state.get(Watched.Configuration))
+    }
+
+    // sbt 1.1.4 and earlier uses Watched.ContinuousState while watching the file system.
+    def watchUsingContinuousState = {
+      // If we have both Watched.Configuration and Watched.ContinuousState
+      // attributes and if Watched.ContinuousState.count is 1 then we assume
+      // we're in ~ run mode
+      for {
+        watched    <- state.get(Watched.Configuration)
+        watchState <- state.get(Watched.ContinuousState)
+        if watchState.count == 1
+      } yield watched
+    }
+
+    val _ :: minor :: patch :: Nil = sbtVersion.split("\\.").map(_.takeWhile(_.isDigit).toInt).toList
+
+    if (minor >= 2) { // sbt 1.2.x and later
+      watchUsingEvenMonitor
+    } else if (minor == 1 && patch >= 5) { // sbt 1.1.5+
+      watchUsingEvenMonitor
+    } else { // sbt 1.1.4 and earlier
+      watchUsingContinuousState
     }
   }
