+++ dev-mode/sbt-scripted-tools/src/main/scala/play/sbt/scriptedtools/ScriptedTools.scala
@@ -28,7 +28,17 @@ import play.sbt.routes.RoutesCompiler.autoImport._
 import play.sbt.run.PlayRun
 import play.sbt.run.toLoggerProxy
 
-object ScriptedTools extends AutoPlugin with ScriptedTools0 {
+import scala.reflect.ClassTag
+import scala.reflect.classTag
+
+import org.apache.logging.log4j.Level
+import org.apache.logging.log4j.core.{ LogEvent => Log4JLogEvent, _ }
+import org.apache.logging.log4j.core.Filter.Result
+import org.apache.logging.log4j.core.appender.AbstractAppender
+import org.apache.logging.log4j.core.filter.LevelRangeFilter
+import org.apache.logging.log4j.core.layout.PatternLayout
+
+object ScriptedTools extends AutoPlugin {
   override def requires = plugins.JvmPlugin
   override def trigger  = allRequirements
 
@@ -209,6 +219,30 @@ object ScriptedTools extends AutoPlugin with ScriptedTools0 {
     sourceLines.foreach { sl =>
       if (!targetLines.contains(sl)) {
         throw new RuntimeException(s"File $target didn't contain line:\n$sl")
+      }
+    }
+  }
+
+  def assertNotEmpty[T: ClassTag](o: java.util.Optional[T]): T = {
+    if (o.isPresent) o.get()
+    else throw new Exception(s"Expected Some[${classTag[T]}]")
+  }
+
+  def bufferLoggerMessages = bufferLogger.messages
+
+  // sbt 1.0 defines extraLogs as a SettingKey[ScopedKey[_] => Seq[Appender]]
+  // while sbt 0.13 uses SettingKey[ScopedKey[_] => Seq[AbstractLogger]]
+  object bufferLogger
+      extends AbstractAppender(
+        "FakeAppender",
+        LevelRangeFilter.createFilter(Level.ERROR, Level.ERROR, Result.NEUTRAL, Result.DENY),
+        PatternLayout.createDefaultLayout()
+      ) {
+    @volatile var messages = List.empty[String]
+
+    override def append(event: Log4JLogEvent): Unit = {
+      if (event.getLevel == Level.ERROR) synchronized {
+        messages = event.getMessage.getFormattedMessage :: messages
       }
     }
   }
